Disable sqrt-table in pasta_curves, since it requires A extension
Note even when we have sqrt-table enable, lazy_static means it will
build the sqrt-table when booting. This is not suitable for ckb's
environment. It's better to employ the same trick as secp256k1: we
build the sqrt table externally, then load the full table directly to
save running time.